auditor/requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
idna==3.4
markdown-it-py==3.0.0
mdurl==0.1.2
neo4j==5.13.0
numpy==1.26.1
pandas==2.1.1
Pygments==2.16.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
rich==13.6.0
six==1.16.0
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
websocket-client==1.6.4
